refactor: 하나에 Service와 Controller에서 사용되었던 기능들을 각 패키지로 분리
