+++ build.gradle
@@ -24,7 +24,7 @@ dependencies {
     implementation 'io.github.cdimascio:dotenv-java:3.0.0'
 
     // AWS S3
-//    implementation 'io.awspring.cloud:spring-cloud-aws-starter-s3:4.0.0'
+    implementation 'io.awspring.cloud:spring-cloud-aws-starter-s3:4.0.0'
 //
     implementation 'org.springframework.boot:spring-boot-starter-validation'
