+++ src/main/java/com/example/i_commerce/domain/chat/repository/ChatRoomRepository.java
@@ -1,6 +1,7 @@
 package com.example.i_commerce.domain.chat.repository;
 
 import com.example.i_commerce.domain.chat.entity.ChatRoom;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,5 +9,6 @@ import org.springframework.stereotype.Repository;
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
 
     boolean existsByProductIdAndIsGroupChat(Long productId, Boolean isGroupChat);
+    List<ChatRoom> findByIsGroupChatTrueAndDeletedAtIsNull();
 
 }
