+++ src/main/java/com/example/i_commerce/domain/chat/service/dto/GroupChatListResponse.java
@@ -1,0 +1,8 @@
+package com.example.i_commerce.domain.chat.service.dto;
+
+public record GroupChatListResponse(
+    Long roomId,
+    String roomName
+) {
+
+}
