+++ src/main/resources/application-dev.yaml
@@ -28,28 +28,22 @@ spring:
     show-sql: true
     open-in-view: false
 
-#  cloud:
-#    aws:
-#      credentials:
-#        access-key: ${AWS_ACCESS_KEY}
-#        secret-key: ${AWS_SECRET_KEY}
-#      region:
-#        static: ap-northeast-2
-#      s3:
-#        bucket: ${AWS_S3_BUCKET_NAME}
-#      stack:
-#        auto: false
+  cloud:
+    aws:
+      credentials:
+        access-key: ${AWS_ACCESS_KEY}
+        secret-key: ${AWS_SECRET_KEY}
+      region:
+        static: ap-northeast-2
+      s3:
+        bucket: ${AWS_S3_BUCKET_NAME}
+      stack:
+        auto: false
 
   servlet:
     multipart:
       max-file-size: 10MB
       max-request-size: 20MB
-
-#임시적인 JWT 설정 추후 필수 삭제
-jwt:
-  # chatserversecretaccesstokenchatserversecretaccesstokenchatserversecretaccesstoken
-  secretKey: Y2hhdHNlcnZlcnNlY3JldGFjY2Vzc3Rva2VuY2hhdHNlcnZlcnNlY3JldGFjY2Vzc3Rva2VuY2hhdHNlcnZlcnNlY3JldGFjY2Vzc3Rva2Vu
-  expiration: 3000 # 3000?
 
 toss:
   secretKey: 1234
